+++ rpki-validator/src/main/java/net/ripe/rpki/validator3/api/roaprefixassertions/RoaPrefixAssertionsController.java
@@ -82,8 +82,8 @@ public class RoaPrefixAssertionsController {
 
     @GetMapping
     public ResponseEntity<ApiResponse<Stream<RoaPrefixAssertionResource>>> list(
-        @RequestParam(name = "startFrom", defaultValue = "0") int startFrom,
-        @RequestParam(name = "pageSize", defaultValue = "20") int pageSize,
+        @RequestParam(name = "startFrom", defaultValue = "0") long startFrom,
+        @RequestParam(name = "pageSize", defaultValue = "20") long pageSize,
         @RequestParam(name = "search", defaultValue = "", required = false) String searchString,
         @RequestParam(name = "sortBy", defaultValue = "prefix") String sortBy,
         @RequestParam(name = "sortDirection", defaultValue = "asc") String sortDirection) {
@@ -94,7 +94,7 @@ public class RoaPrefixAssertionsController {
 
         final List<RoaPrefixAssertion> matching = roaPrefixAssertions.find(searchTerm, sorting, paging).collect(Collectors.toList());
 
-        int totalSize = (int) roaPrefixAssertions.count(searchTerm);
+        long totalSize = (int) roaPrefixAssertions.count(searchTerm);
 
         final Links links = Paging.links(
             startFrom, pageSize, totalSize,
